Implement persistent header storage using sled
