+++ node/Cargo.toml
@@ -28,6 +28,7 @@ tokio = { version = "1.32.0", features = ["macros", "sync"] }
 tracing = "0.1.37"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+sled = "0.34" # upgrade requires migration https://docs.rs/sled/latest/sled/struct.Db.html#examples-1
 tokio = { version = "1.32.0", features = ["rt-multi-thread"] }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
