+++ node/tests/node.rs
@@ -4,7 +4,7 @@ use std::time::Duration;
 use celestia_node::{
     node::{Node, NodeConfig},
     p2p::P2pService,
-    store::InMemoryStore,
+    store::in_memory_store::InMemoryStore,
     test_utils::{test_node_config, test_node_config_with_keypair},
 };
 use celestia_rpc::prelude::*;
